// apps/backend/src/modules/orders/domain/order-status.enum.ts
export enum OrderStatus {
  DRAFT = "draft",
  PENDING_CONFIRMATION = "pending_confirmation",
  PENDING_PAYMENT = "pending_payment",
  PAID = "paid",
  PROCESSING = "processing",
  READY_FOR_SHIPMENT = "ready_for_shipment",
  SHIPPED = "shipped",
  DELIVERED = "delivered",
  COMPLETED = "completed",
  CANCELLED = "cancelled",
  REFUNDED = "refunded",
  PARTIALLY_REFUNDED = "partially_refunded",
}